# examples/swaggercodegen/api/resources/swagger.yaml
swagger: "2.0"
info:
  description: "This service provider allows the creation of fake 'cdns' and 'lbs' resources"
  version: "1.0.0"
  title: "Dummy Service Provider generated using 'swaggercodegen' that has two resources 'cdns' and 'lbs' which are terraform compliant"
  contact:
    email: "apiteam@serviceprovider.io"
#host: "localhost:8443" # If host is not specified, it is assumed to be the same host where the API documentation is being served.
#basePath: ""
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
  - apikey_auth: []

# This make the provider multiregional, so API calls will be make against the specific region as per the value provided
# provided by the user according to the 'x-terraform-provider-regions' regions. If non is provided, the default value will
# be the first item in the 'x-terraform-provider-regions' list of strings. in the case below that will be 'rst1'
...
paths:
...
  /v1/cdns:
    post:
      x-terraform-resource-name: "cdn"
      x-terraform-resource-host: localhost:8443 # If this extension is specified, it will override the global host and API calls will be made against this host instead
      tags:
      - "cdn"
      summary: "Create cdn"
      operationId: "ContentDeliveryNetworkCreateV1"
      parameters:
      - in: "body"
        name: "body"
        description: "Created CDN"
        required: true
        schema:
          $ref: "#/definitions/ContentDeliveryNetworkV1"
      - in: "header"
        x-terraform-header: x_request_id
        name: "X-Request-ID"
        type: "string"
        required: true
      x-terraform-resource-timeout: "30s"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ContentDeliveryNetworkV1"
        default:
          description: "generic error response"
          schema:
            $ref: "#/definitions/Error"
      #security: For the sake of the example, this POST operation will use the global security schemes
      #  - apikey_auth: []
  /v1/cdns/{id}:
    get:
      tags:
      - "cdn"
      summary: "Get cdn by id"
      description: ""
      operationId: "ContentDeliveryNetworkGetV1"
      parameters:
      - name: "id"
        in: "path"
        description: "The cdn id that needs to be fetched."
        required: true
        type: "string"
      #x-terraform-resource-timeout: "30s" If a given operation does not have the 'x-terraform-resource-timeout' extension; the resource operation timeout will default to 10m (10 minutes)
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ContentDeliveryNetworkV1"
        400:
          description: "Invalid cdn id supplied"
        404:
          description: "CDN not found"
      security:
        - apikey_auth: []
    put:
      tags:
      - "cdn"
      summary: "Updated cdn"
      operationId: "ContentDeliveryNetworkUpdateV1"
      parameters:
      - name: "id"
        in: "path"
        description: "cdn that needs to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated cdn object"
        required: true
        schema:
          $ref: "#/definitions/ContentDeliveryNetworkV1"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ContentDeliveryNetworkV1"
        400:
          description: "Invalid cdn id supplied"
        404:
          description: "CDN not found"
      security:
        - apikey_auth: []
    delete:
      tags:
      - "cdn"
      summary: "Delete cdn"
      operationId: "ContentDeliveryNetworkDeleteV1"
      parameters:
      - name: "id"
        in: "path"
        description: "The cdn that needs to be deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "successful operation, no content is returned"
        400:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
      security:
        - apikey_auth: []

  ######################
  ##### LB Resource ####
  ######################
...
securityDefinitions:
  apikey_auth: # basic apikey header auth, the header name will be the value used in the name property, in this case "Authorization", when calling the applicable resource API
    type: "apiKey"
    name: "Authorization"
    in: "header"
#  apikey_header_auth_bearer: // example of header auth using the bearer schema as per the specification
#    type: "apiKey"
#    in: "header"
#    x-terraform-authentication-scheme-bearer: true # this extension would make the auth use the Bearer schema, so
# there will be no need to specify the name property as internally the provider will take care of using the right header name
# following the bearer spec, hence it will use as header name "Authorization", and the token value will be prefixed with
# Bearer schema automatically without needing that input from the user, just the token
#  apikey_query_auth_bearer:  // example of query auth using the bearer schema as per the specification
#    type: "apiKey"
#    in: "query"
#    x-terraform-authentication-scheme-bearer: true # this extension would make the auth use the Bearer schema, so
# there will be no need to specify the name property as internally the provider will take care of using the right header name
# following the bearer spec, hence it will use as header name "Authorization", and the token value will be the one provided
# by the user
#  apikey_query_auth: // basic apikey query auth, the call to the API will attach to the URI the query param, e,g: http://hostname.com?Authorization="value provided by the user"
#    type: "apiKey"
#    name: "Authorization"
#    in: "query"
definitions:
  ContentDeliveryNetworkV1:
    type: "object"
    required:
      - label
      - ips
      - hostnames
    properties:
      id:
        type: "string"
        readOnly: true # This property will not be considered when creating a new resource, however, it is expected to
                       # to be returned from the api, and will be saved as computed value in the terraform state file
      label:
        type: "string"
        x-terraform-immutable: true
      ips:
        type: "array"
        x-terraform-force-new: true # when this value changes terraform will force the creation of a new resource
        items:
          type: "string"
      hostnames:
        type: "array"
        items:
          type: "string"
      exampleInt: # this name is not terraform name compliant; the provider will do a translation on the fly to make it terraform name compliant - the result will be example_int
        type: integer
      exampleNumber:
        type: number
        x-terraform-field-name: betterExampleNumberFieldName  # overriding exampleNumber with a different name 'betterExampleNumberFieldName'; the preferred name is not terraform compliant either so the provider will perform the name conversion automatically when translating the name into the provider resource configuration and when saving the field into the state file
      example_boolean:
        type: boolean
      object_property:
        #type: object - type is optional for properties of object type that use $ref
        $ref: "#/definitions/ObjectProperty"
      arrayOfObjectsExample: # This is an example of an array of objects
        type: "array"
        items:
          type: "object"
          properties:
            protocol:
              type: string
            originPort:
              type: integer
              x-terraform-field-name: "origin_port"
      object_nested_scheme_property: # this also covers object within objects
        type: "array" # as per 'apparentlymart' comment, this is the only way to make nested structs possible in terraform at the moment
        x-terraform-nested-object: true
        items:
          type: "object"
          required:
          - object_property
          properties:
            name:
              type: "string"
              readOnly: true
            object_property:
              type: "array"
              x-terraform-nested-object: true
              items:
                type: "object" # nested properties required type equal object to be considered as object
                properties:
                  account:
                    type: string
  ObjectProperty:
    type: object
    required:
    - message
    - detailedMessage
    - exampleInt
    - exampleNumber
    - example_boolean
    properties:
      message:
        type: string
      detailedMessage:
        type: string
        x-terraform-field-name: "detailed_message"
      exampleInt:
        type: integer
      exampleNumber:
        type: number
      example_boolean:
        type: boolean
...
  Error:
    type: object
    required:
    - code
    - message
    properties:
      code:
        type: string
      message:
        type: string

# Descriptions of common responses
responses:
  NotFound:
    description: The specified resource was not found
    schema:
      $ref: "#/definitions/Error"
  Unauthorized:
    description: Unauthorized
    schema:
      $ref: "#/definitions/Error"